% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Sectional curvature of a rstan object for convergence diagnosis.}
\usage{
curvature(fit)
}
\description{
Sectional curvature of a rstan object for convergence diagnosis.
}
