% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_curvature.R
\name{sectional_curvature}
\alias{sectional_curvature}
\title{Internal function to compute sectional curvature from Hamiltonian Monte Carlo samples.}
\usage{
sectional_curvature(gradU, hessU, Kp)
}
\description{
Internal function to compute sectional curvature from Hamiltonian Monte Carlo samples.
}
