% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Sectional curvature of a rstan object for convergence diagnosis.}
\usage{
curvature(fit, par = "y", num_secs = 1000, num_cores = 4)
}
\description{
Sectional curvature of a rstan object for convergence diagnosis.
}
